+++ machine-learning-client/dockerfile
@@ -8,6 +8,12 @@ WORKDIR /app
 COPY requirements.txt ./
 RUN pip3 install -r requirements.txt
 
+# install tesseract
+RUN apt-get update && apt-get install -y \
+    tesseract-ocr \
+    libtesseract-dev \
+    libgl1-mesa-glx
+
 # the ADD command is how you add files from your local machine into a Docker image
 # Copy the current directory contents into the container at /app
 ADD . .
@@ -16,4 +22,4 @@ ADD . .
 EXPOSE 8000
 
 # Run app.py when the container launches
-CMD [ "python3", "app.py"]
+CMD [ "python3", "client.py"]
