+++ machine-learning-client/requirements.txt
@@ -1,0 +1,30 @@
+astroid==3.3.9
+black==25.1.0
+blinker==1.9.0
+certifi==2025.1.31
+click==8.1.8
+dill==0.3.9
+distlib==0.3.9
+filelock==3.18.0
+Flask==3.1.0
+isort==6.0.1
+itsdangerous==2.2.0
+Jinja2==3.1.6
+MarkupSafe==3.0.2
+mccabe==0.7.0
+mypy-extensions==1.0.0
+numpy==2.2.4
+opencv-python==4.11.0.86
+packaging==24.2
+pathspec==0.12.1
+pillow==11.1.0
+pipenv==2024.4.1
+platformdirs==4.3.7
+pylint==3.3.6
+pytesseract==0.3.13
+setuptools==78.1.0
+tomli==2.2.1
+tomlkit==0.13.2
+typing_extensions==4.13.1
+virtualenv==20.30.0
+Werkzeug==3.1.3
